requirements.txt
asyncgui==0.6.3
asynckivy==0.6.4
certifi==2024.8.30
charset-normalizer==3.4.0
docutils==0.21.2
idna==3.10
Kivy==2.3.0
Kivy-Garden==0.1.5
kivymd==1.2.0
materialyoucolor==2.0.9
pillow==11.0.0
Pygments==2.18.0
PyMuPDF==1.24.13
requests==2.32.3
urllib3==2.2.3
